backend/requirements.txt
# Jiwar Backend - Dependencies (Python 3.13 compatible)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.25
alembic>=1.13.1
psycopg[binary]>=3.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic[email]>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
slowapi>=0.1.9
firebase-admin>=6.5.0
